requirements.txt
torch>=2.0.0
transformers>=4.30.0
Pillow>=8.2.0
Flask>=2.0.0